Rename edit_handlers into panels to match Wagtail's naming
