+++ wagtailgmaps/panels.py
